pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-organizer"
version = "1.0.0"
description = "AI-powered email management system that automatically categorizes and organizes emails across multiple Gmail accounts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
]
keywords = ["gmail", "email", "ai", "organizer", "classification", "claude"]
dependencies = [
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.116.0",
    "anthropic>=0.76.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.31.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/dmhernandez2525/gmail-organizer"
Documentation = "https://gmail-organizer-site.onrender.com"
Repository = "https://github.com/dmhernandez2525/gmail-organizer"

[project.scripts]
gmail-organizer = "gmail_organizer.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gmail_organizer*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
